Created base structure for project
Created the files that will be used to complete this project. As they
may not be final, they are my first thoughts for how it should be
constructed to succeed. The src directory contains all source code
files. The discord directory contains all files pertaining to
manipulating the bot. The game directory contains all files pertaining
to playing a game of tic-tac-toe.