Implemented some input checking and game
Cleaned up a minor bug that would cause the game to break. Also, have
implemented more input checking for the coordinates. It seems that a
game between two users on a server can be done. I have made it so only
one game can be played at a time for less confusion and less text on the
server. I will be taking a break from this project for now, as it is in
a pretty good state and will like to focus on something else. Will come
back again to implement more things like database and VPS.